fix(suite): remove circular dependency in CustomFee
- Extracted CustomFeeBasicProps type to separate types.ts file
- CustomFeeEthereum now imports type directly instead of from CustomFee
- This breaks the circular import between CustomFee and CustomFeeEthereum
- Resolves: packages/suite/src/components/wallet/Fees/CollapsibleFees/CustomFee/CustomFee.tsx > packages/suite/src/components/wallet/Fees/CollapsibleFees/CustomFee/CustomFeeEthereum.tsx
