+++ packages/suite/src/components/wallet/Fees/CollapsibleFees/CustomFee/CustomFee.tsx
@@ -1,4 +1,4 @@
-import { Translation, TranslationFunction, useTranslation } from '@suite/intl';
+import { Translation, useTranslation } from '@suite/intl';
 import { getFeeUnits, isInteger } from '@suite-common/wallet-utils';
 import { Column, Row, Text } from '@trezor/components';
 import { spacings } from '@trezor/theme';
@@ -7,17 +7,8 @@ import { CurrentFee } from './CurrentFee';
 import { CustomFeeEthereum } from './CustomFeeEthereum';
 import { CustomFeeMisc } from './CustomFeeMisc';
 import { CustomFeeTooLowBanner } from './CustomFeeTooLowBanner';
+import type { CustomFeeBasicProps } from './types';
 import { useFeesContext } from '../../context/FeesContext';
-
-export type CustomFeeBasicProps = {
-    composedFeePerByte: string | undefined;
-    translationString: TranslationFunction;
-    feeUnits: string;
-    sharedRules: {
-        required: string;
-        validate: (value: string) => string | undefined;
-    };
-};
 
 type CustomFeeProps = {
     showCurrentFee: boolean;
@@ -81,3 +72,5 @@ export const CustomFee = ({ showCurrentFee }: CustomFeeProps) => {
         </>
     );
 };
+
+export type { CustomFeeBasicProps };
