+++ packages/suite/src/components/wallet/Fees/CollapsibleFees/CustomFee/types.ts
@@ -1,0 +1,11 @@
+import { TranslationFunction } from '@suite/intl';
+
+export type CustomFeeBasicProps = {
+    composedFeePerByte: string | undefined;
+    translationString: TranslationFunction;
+    feeUnits: string;
+    sharedRules: {
+        required: string;
+        validate: (value: string) => string | undefined;
+    };
+};
